requirements.txt
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.3.0
postgrest>=0.13.0
matplotlib>=3.8.0
mplfinance>=0.12.10b0
psycopg2-binary>=2.9.9
packaging>=23.0
